DirectX12: Implement pipeline/buffer binding, format fixes
- Implement BindPipelineImpl, BindVertexBufferImpl, and BindIndexBufferImpl in DXCommandBuffer for correct DX12 usage.
- Pass scissors and viewports by ref to match DX12 API.
- Update DXFormats: return both PrimitiveTopologyType and D3DPrimitiveTopology; add IndexFormat conversion; refactor CommandQueueType mapping.
- Use new primitive topology tuple in DXGraphicsPipeline.
- Change RenderStates.StencilReference to uint for type safety.
- Use unsigned StencilReference in Vulkan pipelines.
- Add missing using directive for Silk.NET.Core.Native.

+++ sources/Zenith.NET.Vulkan/VKGraphicsPipeline.cs
@@ -85,7 +85,7 @@ internal unsafe class VKGraphicsPipeline : GraphicsPipeline
                     CompareOp = VKFormats.Vulkan(desc.RenderStates.DepthStencilState.FrontFace.StencilFunc),
                     CompareMask = desc.RenderStates.DepthStencilState.StencilReadMask,
                     WriteMask = desc.RenderStates.DepthStencilState.StencilWriteMask,
-                    Reference = (uint)desc.RenderStates.StencilReference
+                    Reference = desc.RenderStates.StencilReference
                 },
                 Back = new()
                 {
@@ -95,7 +95,7 @@ internal unsafe class VKGraphicsPipeline : GraphicsPipeline
                     CompareOp = VKFormats.Vulkan(desc.RenderStates.DepthStencilState.BackFace.StencilFunc),
                     CompareMask = desc.RenderStates.DepthStencilState.StencilReadMask,
                     WriteMask = desc.RenderStates.DepthStencilState.StencilWriteMask,
-                    Reference = (uint)desc.RenderStates.StencilReference
+                    Reference = desc.RenderStates.StencilReference
                 },
                 MinDepthBounds = 0.0f,
                 MaxDepthBounds = 1.0f
